Resolve inconsistencies with valve statuses in disarm()
Follow on of 987777d10436639d3124edaab32d6d8bb942d07a

Another hunch, I think we want the vent valve open, to relieve pressure in
fuel feed system.

Assuming, just based on my previous status reasoning, that we want
passivation of the entire system, therefore we want to vent pressure in
the ACS as well. However, the big issue here is that the only way to
vent the ACS is to turn on the thrusters, which sounds a bit less
ideal, like a flopping fish (more potential damage and damage of
rupture, etc.)