Change accel and gyro references to get data from MPU
There are a couple things that could be modified/made more robust, for
example the ACCEL_SENSITIVITY and GYRO_SENSITIVITY constants are
hardcoded, yet it is possbile to read what these values are set to from
the MPU itself, so if it is configured differently, the sensitivity can
adapt dynamically.

This code compiles, and in theory it runs, however it has had issues
with the latter as long as I've had a hold of it. I haven't changed
anything to affect how it runs.